# colexecjoin: optimize building output on the left in merge joiner