--- a/pkg/col/coldata/vec_tmpl.go
+++ b/pkg/col/coldata/vec_tmpl.go
@@ -210,7 +210,7 @@ func (m *memColumn) Copy(args SliceArgs) {
 							m.nulls.SetNull(i + args.DestIdx)
 						} else {
 							// {{if .IsBytesLike}}
-							toCol.copy(fromCol, i+args.DestIdx, selIdx)
+							toCol.Copy(fromCol, i+args.DestIdx, selIdx)
 							// {{else}}
 							v := fromCol.Get(selIdx)
 							// {{if .Sliceable}}
@@ -238,7 +238,7 @@ func (m *memColumn) Copy(args SliceArgs) {
 					//gcassert:bce
 					selIdx := sel[i]
 					// {{if .IsBytesLike}}
-					toCol.copy(fromCol, i+args.DestIdx, selIdx)
+					toCol.Copy(fromCol, i+args.DestIdx, selIdx)
 					// {{else}}
 					v := fromCol.Get(selIdx)
 					// {{if .Sliceable}}
@@ -285,7 +285,7 @@ func _COPY_WITH_REORDERED_SOURCE(_SRC_HAS_NULLS bool) { // */}}
 		// {{end}}
 		{
 			// {{if .IsBytesLike}}
-			toCol.copy(fromCol, destIdx, srcIdx)
+			toCol.Copy(fromCol, destIdx, srcIdx)
 			// {{else}}
 			v := fromCol.Get(srcIdx)
 			toCol.Set(destIdx, v)