… pkg/sql/colexec/colexecjoin/mergejoiner_exceptall.eg.go
… pkg/sql/colexec/colexecjoin/mergejoiner_fullouter.eg.go
… pkg/sql/colexec/colexecjoin/mergejoiner_inner.eg.go
… pkg/sql/colexec/colexecjoin/mergejoiner_intersectall.eg.go
… pkg/sql/colexec/colexecjoin/mergejoiner_leftanti.eg.go
… pkg/sql/colexec/colexecjoin/mergejoiner_leftouter.eg.go
… pkg/sql/colexec/colexecjoin/mergejoiner_leftsemi.eg.go
… pkg/sql/colexec/colexecjoin/mergejoiner_rightanti.eg.go
… pkg/sql/colexec/colexecjoin/mergejoiner_rightouter.eg.go
… pkg/sql/colexec/colexecjoin/mergejoiner_rightsemi.eg.go